Move node specific DER stuff to crypto.js.
